+++ src/ClassFinder.php
@@ -1,6 +1,9 @@
 <?php
 namespace HaydenPierce\ClassFinder;
 
+use HaydenPierce\ClassFinder\Classmap\ClassmapEntryFactory;
+use HaydenPierce\ClassFinder\Classmap\ClassmapFinder;
+use HaydenPierce\ClassFinder\Exception\ClassFinderException;
 use HaydenPierce\ClassFinder\PSR4\PSR4Finder;
 use HaydenPierce\ClassFinder\PSR4\PSR4NamespaceFactory;
 
@@ -12,6 +15,9 @@ class ClassFinder
     /** @var PSR4Finder */
     private static $psr4;
 
+    /** @var ClassmapFinder */
+    private static $classmap;
+
     private static function initialize()
     {
         if (!(self::$config instanceof AppConfig)) {
@@ -21,6 +27,11 @@ class ClassFinder
         if (!(self::$psr4 instanceof PSR4Finder)) {
             $PSR4Factory = new PSR4NamespaceFactory(self::$config);
             self::$psr4 = new PSR4Finder($PSR4Factory);
+        }
+
+        if (!(self::$classmap instanceof ClassmapFinder)) {
+            $classmapFactory = new ClassmapEntryFactory(self::$config);
+            self::$classmap = new ClassmapFinder($classmapFactory);
         }
     }
 
@@ -33,9 +44,20 @@ class ClassFinder
     {
         self::initialize();
 
-        $classes = self::$psr4->findClasses($namespace);
+        $isNamespaceKnown = self::$psr4->isNamespaceKnown($namespace) || self::$classmap->isNamespaceKnown($namespace);
+        if (!$isNamespaceKnown) {
+            throw new ClassFinderException(sprintf("Unknown namespace '%s'. See '%s' for details.",
+                $namespace,
+                'https://gitlab.com/hpierce1102/ClassFinder/blob/master/docs/exceptions/unknownNamespace.md'
+            ));
+        }
 
-        return array_values($classes);
+        $psr4classes = self::$psr4->findClasses($namespace);
+        $classmapClasses = self::$classmap->findClasses($namespace);
+
+        $classes = array_merge($psr4classes, $classmapClasses);
+
+        return array_unique($classes);
     }
 
     public static function setAppRoot($appRoot)