+++ src/Classmap/ClassmapFinder.php
@@ -1,0 +1,45 @@
+<?php
+namespace HaydenPierce\ClassFinder\Classmap;
+
+use HaydenPierce\ClassFinder\FinderInterface;
+
+class ClassmapFinder implements FinderInterface
+{
+    private $factory;
+
+    public function __construct(ClassmapEntryFactory $factory)
+    {
+        $this->factory = $factory;
+    }
+
+    public function isNamespaceKnown($namespace)
+    {
+        $classmapEntries = $this->factory->getClassmapEntries();
+
+        foreach($classmapEntries as $classmapEntry) {
+            if ($classmapEntry->knowsNamespace($namespace)) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    /**
+     * @param $namespace
+     * @return bool|string
+     * @throws ClassFinderException
+     */
+    public function findClasses($namespace)
+    {
+        $classmapEntries = $this->factory->getClassmapEntries();
+
+        $matchingEntries = array_filter($classmapEntries, function(ClassmapEntry $entry) use ($namespace) {
+            return $entry->matches($namespace);
+        });
+
+        return array_map(function(ClassmapEntry $entry) {
+            return $entry->getClassName();
+        }, $matchingEntries);
+    }
+}